services/web/tmtar/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
psycopg2-binary==2.8.4
pytest==6.0.1
flask-marshmallow==0.13.0
flask-restplus==0.13.0
marshmallow-sqlalchemy==0.23.1
Werkzeug==1.0.1
flask-accepts==0.17.3
typing-extensions==3.7.4.2
flask-jwt-extended==3.24.1
marshmallow-enum==1.5.1
gunicorn==20.0.4